scripts/EOS-trainer/requirements.txt
azure-cognitiveservices-speech==1.43.0
azure-cognitiveservices-vision-customvision==3.1.1
azure-common==1.1.28
azure-core==1.32.0
azure-mgmt-core==1.5.0
opencv-python==4.11.0.86
pillow==11.1.0
customtkinter==5.2.2

